fix: fix multistate buttons by moving and adding a few methods and attributes on abstract classes.
